third_party/blink/web_tests/crypto/subtle/derive-hkdf-keys-expected.txt

Test deriving HKDF keys with deriveKey()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Derive an HKDF key from ECDH keys
PASS hkdfKey.algorithm.name is "HKDF"
PASS typeof hkdfKey.extractable is 'boolean'
PASS hkdfKey.extractable is false
PASS hkdfKey.usages.join(',') is "deriveKey,deriveBits"

Derive 128 bits from the HKDF key
PASS bytesToHexString(derivedBits) is "2afc8b03741413cd32a80dc9a8451be8"
PASS successfullyParsed is true

TEST COMPLETE
